Add a version property to ReplicaInfo. This property is used to detect double invocation of ReplicaStorage methods for instance in the case of a backoff.
Simplify the way ReplicaInfo should be used: the previous implementation allowed
the creation of ReplicaInfo instances partially defined in order to send deltas
to ReplicaStorages. This was an implementation trying to avoid to unecessary
serializations. As a matter of fact, it adds complexity and the performance gain
is not significant.