This file was not ci for an unknown reason...
* version PartitionFacade and use this version number to ensure that messages sent from one PartitionFacade to another have the same version.
If the target PartitionFacade has a too low version number - this happens when the target node has not yet processed a partition rebalancing
request - then the processing waits, on the receiving node side, until the PartitionFacade version is updated or until a timeout period. If the
target PartitionFacade has a too high version number - this happens when the sending node has not yet processed a partition rebalancing request
that the receiving node has already processed - then it instructs the requesting node that its version number is too low. Then the sending
PartitionFacade waits until its version is updated or until a timeout.

Thanks to this versioning approach, partitions can be rebalanced when requests are raining on the cluster - as demonstrated by
SimplePartitionManagerSmokeTest.

* carefully review the rebalancing code of SimplePartitionManager: better exception handling; fix re-population problem; and attempt to make the
code more robust.