* Drop Motable.getBodyAsByteBuffer() and Motable.setBodyAsByteBuffer: they were not implemented.
* ClusteredManager takes a ServiceSpace instead of a Dispatcher as constructor
parameter. This is a simple API change to migrate to this new API prior to
a release cut.

* Initialize the ParitionFacades of a SimplePartition when it starts. The
previous implementation was broken as the PartitionFacades were retaining
their state between re-start.

* If a partition transfer is not successful, the PartitionFacades which 
should have been localized are repopulated.

* Fail fast in the case of problems, for instance when a lock cannot be
acquired, by throwing a WADIRuntimeException. A thorough code review is
required to ensure a fail fast behavior.

* Fix ServiceSpaceListener API: add a new parameters, which is the new hosting
peers following the ServiceSpace event. This is required to allow notification
of listeners without having to lock on the hosting peers.

* Drop unused classes.