I have been working offline for nearly 2.5 weeks and I am no more able to commit this work in small pieces...
High-level changes:
* Add Dispatcher.getContext: this returns a DispatcherContext which can be
used to share objects between ServiceSpaces. For instance, a ServiceSpace
adds itself to a ServiceSpaceRegistry bound to this context when it is started.

* Add EnvelopeInterceptor: this is used to intercept and filter Envelopes
as they fly in our out of a Dispatcher.

* Replace RendezVousMsgDispatcher by a more generic and faster one: instead
of using the type of an Envelope payload to identify if we should odd 
it to a RDV, we can simply use a tag against it. This way a single EndPoint
need to be registered to handle 'reply' Envelopes.

* Rename in-here and there Message to Envelope.

* Drop the notion of coordinator from the group API: it is not used and
superseded by the single service capability provided by ServiceSpace. Also,
this simplifies the implementation.

* Add the first cut of an admin ServiceSpace: so far, it allows the discovery
of all the ServiceSpaces running on top of a Dispatcher. More o.c.w.s.a.Command
implementations to follow.

* Drop axis2 code; There are no more interested people in this area.