InstanceTracker is doing too much and some of its responsibilities need to be removed: * InstanceTracker now only tracks updates to an instance; and * it allows clients to retrieve the updates via retrieveValueUpdaterInfos.
Drop TrackedField. After further consideration, it is better to tracked all the field which are not marked as transient so that users are not confused.

Add ClusteredStateMarker.$wadiGetFieldValues() which allows clients to retrieve current field values. This is used to extract the full state of an instance.

Add CopyFullStateVisitor which uses the above new contract to extract and serialize the full state of an instance.