Rewrite replication protocol: now that the implementation is fully sync, I have discovered that it does not scale well when # of peers increases.
The new protocol is less chatty: when a session is relocated from one peer to another peer. The peer owning the session at the beginning of the migration, can becomes a secondary, depending on the backing strategy. And the peer becoming the session owner simply clears its replica storage if it was a secondary.

This new implementation scales independently of the # of nodes. I tested with clusters having from 2 to 5 nodes.